cleanup schemas
